Style - Rename the "pySankey" directory to "pysankey"
Because of "Module name pySankey doesn't conform to snake_case naming
style (invalid-name)"

The change to the package name and to the figure_name attribute makes
this version incompatible with the previous one. According to semantic
versionning we need to increase the version.
