+++ pysankey/__init__.py
@@ -1,6 +1,6 @@
 
 """ Make simple, pretty Sankey Diagrams """
 
-from pySankey.sankey import sankey, PySankeyException, NullsInFrame, LabelMismatch
+from pysankey.sankey import sankey, PySankeyException, NullsInFrame, LabelMismatch
 
 __all__ = ["sankey", "PySankeyException", "NullsInFrame", "LabelMismatch"]
